refactor book services
